+++ tensorrt_llm/_torch/models/modeling_qwen3.py
@@ -5,17 +5,13 @@ import torch
 from torch import nn
 from transformers import PretrainedConfig, Qwen3Config
 
-from tensorrt_llm.functional import PositionEmbeddingType
-
 from ..attention_backend import AttentionMetadata
-from ..attention_backend.interface import PositionalEmbeddingParams, RopeParams
 from ..model_config import ModelConfig
-from ..modules.attention import Attention
 from ..modules.decoder_layer import DecoderLayer
 from ..modules.embedding import Embedding
 from ..modules.gated_mlp import GatedMLP
 from ..modules.linear import TensorParallelMode
-from ..modules.multi_stream_utils import maybe_execute_in_parallel
+from ..modules.qk_norm_attention import QKNormRoPEAttention
 from ..modules.rms_norm import RMSNorm
 from ..speculative import SpecMetadata
 from .modeling_speculative import SpecDecOneEngineForCausalLM
@@ -127,108 +123,6 @@ def compute_yarn_parameters(
     return factor, low, high, attention_factor
 
 
-class Qwen3Attention(Attention):
-
-    def __init__(
-        self,
-        model_config: ModelConfig[Qwen3Config],
-        layer_idx: Optional[int] = None,
-        fuse_qk_norm_rope: bool = True,
-    ):
-        config = model_config.pretrained_config
-        self.pretrained_config = config
-
-        if getattr(config, "rope_scaling", None) is not None:
-            if "type" in config.rope_scaling:
-                pos_type = config.rope_scaling["type"]
-            elif "rope_type" in config.rope_scaling:
-                pos_type = config.rope_scaling["rope_type"]
-            else:
-                raise ValueError(
-                    "rope_scaling must have type or rope_type field")
-            pos_embd_params = PositionalEmbeddingParams(
-                type=PositionEmbeddingType.from_string(pos_type),
-                rope=RopeParams.from_config(config),
-            )
-        else:
-            pos_embd_params = PositionalEmbeddingParams(
-                type=PositionEmbeddingType.rope_gpt_neox,
-                rope=RopeParams.from_config(config),
-            )
-
-        self.fuse_qk_norm_rope = fuse_qk_norm_rope
-
-        super().__init__(
-            hidden_size=config.hidden_size,
-            num_attention_heads=config.num_attention_heads,
-            num_key_value_heads=config.num_key_value_heads,
-            max_position_embeddings=config.max_position_embeddings,
-            bias=config.attention_bias,
-            pos_embd_params=pos_embd_params,
-            rope_fusion=not self.
-            fuse_qk_norm_rope,  # If fuse_qk_norm_rope is true, do not apply fused RoPE in attention OP, and self.rotary_emb will be skipped in the overridden apply_rope.
-            layer_idx=layer_idx,
-            dtype=config.torch_dtype,
-            dense_bias=config.attention_bias,
-            config=model_config,
-        )
-
-        self.q_norm = RMSNorm(hidden_size=self.head_dim,
-                              eps=1e-6,
-                              dtype=config.torch_dtype,
-                              has_weights=True)
-        self.k_norm = RMSNorm(hidden_size=self.head_dim,
-                              eps=1e-6,
-                              dtype=config.torch_dtype,
-                              has_weights=True)
-        self.aux_stream = torch.cuda.Stream()
-        self.ln_events = [torch.cuda.Event(), torch.cuda.Event()]
-
-    def apply_qk_norm(self, q, k):
-
-        def q_l2norm():
-            return self.q_norm(q.reshape(-1, self.head_dim)).reshape(
-                -1, self.q_size)
-
-        def k_l2norm():
-            return self.k_norm(k.reshape(-1, self.head_dim)).reshape(
-                -1, self.kv_size)
-
-        q, k = maybe_execute_in_parallel(
-            q_l2norm,
-            k_l2norm,
-            self.ln_events[0],
-            self.ln_events[1],
-            self.aux_stream,
-        )
-
-        return q, k
-
-    def apply_qk_norm_rope(self, qkv, position_ids):
-        factor, low, high, attention_factor = compute_yarn_parameters(
-            self.pretrained_config)
-        torch.ops.trtllm.fused_qk_norm_rope(
-            qkv, self.num_heads, self.num_key_value_heads,
-            self.num_key_value_heads, self.head_dim,
-            self.q_norm.variance_epsilon, self.q_norm.weight,
-            self.k_norm.weight,
-            self.pos_embd_params.rope.theta, self.pos_embd_params.is_neox,
-            position_ids.view(-1), factor, low, high, attention_factor)
-        return qkv, None, None
-
-    def apply_rope(self, q: torch.Tensor, k: Optional[torch.Tensor],
-                   v: Optional[torch.Tensor], position_ids: torch.Tensor):
-        # Qwen3 applies QK norm before RoPE.
-        if not self.fuse_qk_norm_rope:
-            q, k, v = self.split_qkv(q, k, v)
-            q, k = self.apply_qk_norm(q, k)
-            return super().apply_rope(q, k, v, position_ids)
-
-        assert k is None and v is None, "The input should be a concatenated qkv tensor to apply_qk_norm_rope"
-        qkv = q
-        return self.apply_qk_norm_rope(qkv, position_ids)
-
-
 class Qwen3DecoderLayer(DecoderLayer):
 
     def __init__(
@@ -239,7 +133,7 @@ class Qwen3DecoderLayer(DecoderLayer):
         super().__init__()
         self.layer_idx = layer_idx
         config = model_config.pretrained_config
-        self.self_attn = Qwen3Attention(
+        self.self_attn = QKNormRoPEAttention(
             model_config,
             layer_idx=layer_idx,
         )
